+++ parser/src/test/java/module-info.test
@@ -2,7 +2,7 @@
   org.junitpioneer,org.hamcrest,guava.testlib,creek.test.util
 
 --add-exports
-  creek.system.test.parser/org.creek.system.test.parser.internal=com.fasterxml.jackson.databind
+  creek.system.test.parser/org.creek.internal.system.test.parser=com.fasterxml.jackson.databind
 
 --add-reads
   creek.system.test.parser=org.junitpioneer,org.hamcrest,guava.testlib,creek.test.util
